# lab03/vineward/backend/src/main/java/ch/heig/amt/vineward/api/reference/ReviewRef.java
package ch.heig.amt.vineward.api.reference;

import ch.heig.amt.vineward.business.model.review.ReviewStatus;

public record ReviewRef(
    Long id,
    String title,
    ReviewStatus status,
    UserRef author
) {
}