# lab03/vineward/frontend/openapi/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:13000/api","description":"Generated server url"}],"tags":[{"name":"Moderation","description":"Request moderation"}],"paths":{"/v1/wines/{wineId}/reviews":{"get":{"tags":["wine-review-controller"],"operationId":"getReviews","parameters":[{"name":"wineId","in":"path","required":true,"schema":{"$ref":"#/components/schemas/WineIdAttribute"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReviewViewModel"}}}}}}},"post":{"tags":["wine-review-controller"],"operationId":"postReview","parameters":[{"name":"wineId","in":"path","required":true,"schema":{"$ref":"#/components/schemas/WineIdAttribute"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewViewModel"}}}}}}},"/v1/reviews/{reviewId}/comments":{"get":{"tags":["comment-controller"],"operationId":"getCommentsForReview","parameters":[{"name":"reviewId","in":"path","required":true,"schema":{"$ref":"#/components/schemas/ReviewIdAttribute"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommentViewModel"}}}}}}},"post":{"tags":["comment-controller"],"operationId":"createComment","parameters":[{"name":"reviewId","in":"path","required":true,"schema":{"$ref":"#/components/schemas/ReviewIdAttribute"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentRequestBody"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/v1/moderation/reviews/{reviewId}/submit":{"post":{"tags":["Moderation"],"summary":"Submit a review for moderation","operationId":"submitReview","parameters":[{"name":"reviewId","in":"path","required":true,"schema":{"$ref":"#/components/schemas/ReviewIdAttribute"}}],"responses":{"200":{"description":"OK"}}}},"/v1/moderation/reviews/{reviewId}/deny":{"post":{"tags":["Moderation"],"summary":"Deny the review","operationId":"denyReview","parameters":[{"name":"reviewId","in":"path","required":true,"schema":{"$ref":"#/components/schemas/ReviewIdAttribute"}}],"responses":{"200":{"description":"OK"}}}},"/v1/moderation/reviews/{reviewId}/approve":{"post":{"tags":["Moderation"],"summary":"Approve the review","operationId":"approveReview","parameters":[{"name":"reviewId","in":"path","required":true,"schema":{"$ref":"#/components/schemas/ReviewIdAttribute"}}],"responses":{"200":{"description":"OK"}}}},"/v1/auth/register":{"post":{"tags":["authentication-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequestBody"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/v1/auth/refresh-token":{"post":{"tags":["authentication-controller"],"operationId":"refreshToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationViewModel"}}}}}}},"/v1/auth/login":{"post":{"tags":["authentication-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationViewModel"}}}}}}},"/v1/wines":{"get":{"tags":["wine-controller"],"operationId":"getWines","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WineViewModel"}}}}}}}},"/v1/wines/{id}":{"get":{"tags":["wine-controller"],"operationId":"getWine","parameters":[{"name":"id","in":"path","required":true,"schema":{"$ref":"#/components/schemas/WineIdAttribute"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WineViewModel"}}}}}}},"/v1/reviews":{"get":{"tags":["wine-review-controller"],"operationId":"getReviews_1","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReviewViewModel"}}}}}}}},"/v1/reviews/{reviewId}":{"get":{"tags":["wine-review-controller"],"operationId":"getReview","parameters":[{"name":"reviewId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewViewModel"}}}}}}},"/v1/moderation/reviews/peek":{"get":{"tags":["Moderation"],"summary":"Peek at the first review in the queue without removing it","operationId":"peekNextReview","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PendingReviewModel"}}}}}}},"/v1/moderation/reviews/count":{"get":{"tags":["Moderation"],"operationId":"getPendingReviewModerationCount","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}}}}},"/v1/auth/@me":{"get":{"tags":["authentication-controller"],"operationId":"getUser","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserViewModel"}}}}}}}},"components":{"schemas":{"WineIdAttribute":{"type":"object","properties":{"value":{"type":"integer","format":"int64"}}},"ReviewRequestBody":{"type":"object","properties":{"title":{"type":"string"},"body":{"type":"string"}},"required":["body","title"]},"ReviewViewModel":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"body":{"type":"string"},"author":{"$ref":"#/components/schemas/UserRef"},"wine":{"$ref":"#/components/schemas/WineRef"}},"required":["author","body","id","title","wine"]},"UserRef":{"type":"object","properties":{"displayName":{"type":"string"}}},"WineRef":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"ReviewIdAttribute":{"type":"object","properties":{"value":{"type":"integer","format":"int64"}}},"CommentRequestBody":{"type":"object","properties":{"body":{"type":"string"}},"required":["body"]},"RegisterRequestBody":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"canton":{"type":"string","enum":["AARGAU","APPENZELL_INNERRHODEN","APPENZELL_AUSSERRHODEN","BERN","BASEL_LAND","BASEL_STADT","FRIBOURG","GENEVE","GLARUS","GRAUBUNDEN","JURA","LUZERN","NEUCHATEL","NIDWALDEN","OBWALDEN","SCHAFFHAUSEN","SCHWYZ","SOLOTHURN","ST_GALLEN","THURGAU","TICINO","URI","WALLIS","VAUD","ZUG","ZURICH"]}},"required":["canton","email","password","username"]},"RefreshTokenRequestBody":{"type":"object","properties":{"refreshToken":{"type":"string"}}},"AuthenticationViewModel":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"LoginRequestBody":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"ReviewRef":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"author":{"$ref":"#/components/schemas/UserRef"}}},"WineViewModel":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"origin":{"type":"string"},"region":{"type":"string"},"varieties":{"type":"array","items":{"type":"string"},"uniqueItems":true},"vintner":{"type":"string"},"reviews":{"type":"array","items":{"$ref":"#/components/schemas/ReviewRef"}}},"required":["id","name","origin","region","varieties","vintner"]},"CommentViewModel":{"type":"object","properties":{"body":{"type":"string"},"author":{"$ref":"#/components/schemas/UserRef"}},"required":["author","body"]},"PendingReviewModel":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"body":{"type":"string"},"status":{"type":"string","enum":["DRAFT","PENDING_REVIEW","VERIFIED"]},"wine":{"$ref":"#/components/schemas/WineRef"},"author":{"$ref":"#/components/schemas/UserRef"}}},"UserViewModel":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"roles":{"type":"array","items":{"type":"string","enum":["ADMINISTRATOR","MODERATOR","REVIEWER","USER"]}},"origin":{"type":"string","enum":["AARGAU","APPENZELL_INNERRHODEN","APPENZELL_AUSSERRHODEN","BERN","BASEL_LAND","BASEL_STADT","FRIBOURG","GENEVE","GLARUS","GRAUBUNDEN","JURA","LUZERN","NEUCHATEL","NIDWALDEN","OBWALDEN","SCHAFFHAUSEN","SCHWYZ","SOLOTHURN","ST_GALLEN","THURGAU","TICINO","URI","WALLIS","VAUD","ZUG","ZURICH"]}}}}}}